+++ src/components/Elements/_Portal/ElementRemovePolyfill.js
@@ -1,17 +1,7 @@
-// from:https://github.com/jserz/js_piece/blob/master/DOM/ChildNode/remove()/remove().md
-(function RemovePolyfill(arr) {
-  arr.forEach((item) => {
-    // eslint-disable-next-line no-prototype-builtins
-    if (item.hasOwnProperty('remove')) {
-      return;
+if (!('remove' in Element.prototype)) {
+  Element.prototype.remove = function ElementRemove() {
+    if (this.parentNode) {
+      this.parentNode.removeChild(this);
     }
-    Object.defineProperty(item, 'remove', {
-      configurable: true,
-      enumerable: true,
-      writable: true,
-      value: function remove() {
-        this.parentNode.removeChild(this);
-      },
-    });
-  });
-}([Element.prototype, CharacterData.prototype, DocumentType.prototype]));
+  };
+}